integrated ML with GUI
